fix the bug in getValueInTimestamp of SeqResourceReaderByTimestamp; further split the subpackage chunkRelated into two subpackages chunkRelated and fileRelated.
